// src/models/TitleViewModel.ts
// модель для отображения
// описываю, так как для меня тема новая:
// Данные в БД могут иметь один формат,
// но данные запрашиваемые на фронт (по сути
// view слой) могут иметь другой формат.
// View Model описывает конкретный тип данных
// для отправки.

// Пример: для отправки на view используются
// два значения id и title, однако в БД у
// объектов также есть свойство length,
// которое НЕ описывается представленной моделью,
// т. к. (условно) length не должна
// быть отправлена на фронт (view слой)

export type TitleViewModel = {
    id: number;
    title: string;
};